Some saving functionality
There's definitely a bug problem with line saving, but some of it works.
Look at: fix id number reset, add/remove from allnodes in creator, and
foreach loop: get nodes from ids

+++ MindMap/Assets/Scripts/Nodes/ConnectionSerialized.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 [System.Serializable]
 public class ConnectionSerialized {
 	public string label;
-	public int thickness;
+	public float thickness;
 	public bool isBold;
 	public bool isVisible;
